- fixed dependencies affected by renaming project
