+++ James_Jordan_Portfolio/James_Jordan_Portfolio.http
@@ -1,0 +1,6 @@
+@James_Jordan_Portfolio_HostAddress = http://localhost:5099
+
+GET {{James_Jordan_Portfolio_HostAddress}}/weatherforecast/
+Accept: application/json
+
+###
