--- James_Jordan_Portfolio/James_Jordan_Portfollio.http
@@ -1,6 +1,0 @@
-@James_Jordan_Portfollio_HostAddress = http://localhost:5099
-
-GET {{James_Jordan_Portfollio_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
